Rough cleanup and functionality for single player setup screen
