--- misc/requirements.txt
@@ -1,1 +1,0 @@
-pygame==2.6.1
