+++ pyproject.toml
@@ -1,0 +1,11 @@
+[project]
+name = "codingbros"
+version = "0.1.0"
+description = "Add your description here"
+readme = "README.md"
+requires-python = ">=3.13"
+dependencies = [
+    "pygame==2.6.1",
+    "pygame-popup>=0.10.0",
+    "setuptools>=80.9.0",
+]
