+++ src/button/checkbox.py
@@ -1,8 +1,33 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+
+"""
+File: checkbox.py
+Description: This is the file containing the code for custom checkboxes used on different setup screens
+
+Author: Coding Bros
+Email: See team info <project_root>/misc/teamInfo.txt
+Date Created: 2024-10-13
+Last Modified: 2024-10-31
+Version: 1.0
+"""
+
+
+# ===============
+# Start Imports
+# ===============
 import pygame as pg
+
+# ===============
+# End Imports
+# ===============
+
+
+# Pygame initialization
 pg.font.init()
 
 class Checkbox:
-    def __init__(self, surface, x, y, idnum, color=(230, 230, 230),
+    def __init__(self, surface, x, y, idnum, type, color=(230, 230, 230),
         caption="", outline_color=(0, 0, 0), check_color=(255,180,0),
         font_color=(0, 0, 0), text_offset=(35, 1), font=pg.font.Font(None,16)):
         self.surface = surface
@@ -15,6 +40,7 @@ class Checkbox:
         self.fc = font_color
         self.to = text_offset
         self.ft = font
+        self.type = type
 
         #identification for removal and reorginazation
         self.idnum = idnum
@@ -45,7 +71,7 @@ class Checkbox:
             pg.draw.rect(surface, self.oc, self.checkbox_outline, 1)
         self._draw_button_text(surface)
 
-    def _update(self, event, ):
+    def _update(self, event):
         x, y = pg.mouse.get_pos()
         px, py, w, h = self.checkbox_obj
         if px < x < px + w and py < y < py + w:
@@ -57,10 +83,14 @@ class Checkbox:
     def update(self, surface):
         self.render_checkbox(surface)
 
-    def on_click(self,event):
+    def on_click(self, event):
         self._update(event)
 
     def check_event(self, event):
         """The button needs to be passed events from your program event loop."""
         if event.type == pg.MOUSEBUTTONDOWN and event.button == 1:
             self.on_click(event)
+
+    def reset_to_default(self, surface):
+        self.checked = False
+        self.update(surface)
