+++ src/gamestate.py
@@ -1,0 +1,64 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+
+"""
+File: gamestate.py
+Description: This is the parent class for individual states/screens to inherit from
+
+Author: Coding Bros
+Email: See team info <project_root>/misc/teamInfo.txt
+Date Created: 2024-10-13
+Last Modified: 2024-10-31
+Version: 1.0
+"""
+# ===============
+# Start Imports
+# ===============
+import sys
+import pygame as pg
+
+# ===============
+# End Imports
+# ===============
+
+class GameState(object):
+    """
+    Parent class for individual game states to inherit from. 
+    """
+    def __init__(self):
+        self.done = False
+        self.quit = False
+        self.next_state = None
+        self.screen_rect = pg.display.get_surface().get_rect()
+        self.persist = {}
+        self.font = pg.font.Font(None, 24)
+
+    def startup(self, persistent):
+        """
+        Called when a state resumes being active.
+        Allows information to be passed between states.
+        
+        persistent: A dict passed from state to state
+        """
+        self.persist = persistent        
+        
+    def get_event(self, event):
+        """
+        Handle a single event passed by the Game object.
+        """
+        pass
+        
+    def update(self, dt):
+        """
+        Update the state. Called by the Game object once
+        per frame. 
+        
+        dt: time since last frame
+        """
+        pass
+        
+    def draw(self, surface):
+        """
+        Draw everything to the screen.
+        """
+        pass
