+++ src/globals.py
@@ -1,0 +1,25 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+
+"""
+File: globals.py
+Description: This is the file contain globals/constants used across all TicTactoe Screens
+"""
+
+class TicTacToeGlobals:
+    """
+    Stores info used by all classes for this project
+    """
+    # Dimensions
+    WIDTH = 600
+
+    # Colors
+    BLACK = (0,0,0)
+    WHITE = (255,255,255)
+    ORANGE = (255,180,0)
+    GREEN = (106,168,79)
+    BG_COLOR = (67,67,67)
+
+    # File Paths
+    EIGHT_BIT_FONT_PATH = "font/Eight-Bit Madness.ttf"
+    BLIP_SHORT_SOUND_PATH = "sounds/blipshort1.wav"
